refs #140: new optimization: insert clr to reduce and #255
before each assignment of a byte or word from mem into a previously
unused register, clear that register.

This helps the existing optimizations to eliminate 'and' for unsigned
extends. The 'clr' insns without effect are also removed.

This leads to better code: http://franke.ms/cex/z/8xecqW

Also fixed a bug in the usage analyzer.